feat: ✨  add test file for version steps (#29)
* feat: add dataset version test

* fix: confest.py password change

* style: 🎨 fix code style issues with Black

* fix: dataset & its metadata http request status code

* fix: study metadata http statuses fixed

* fix: study status test

* style: 🎨 fix code style issues with Black

* fix: password was reverted

* fix: delete endpoint status

* fix: delete endpoints

* fix: study dataset test

* fix: minor dataset typo fixes

* fix: delete endpoints

* fix: dataset relate item test functions

* style: 🎨 fix code style issues with Black

* fix: restored some study metadata fields

* fix: study version minimised endpoint

* style: format

* style: 🎨 fix code style issues with Black

---------

+++ tests/functional/test_study_dataset_metadata_api.py
@@ -87,7 +87,7 @@ def test_post_alternative_identifier(_logged_in_client):
         ],
     )
 
-    assert response.status_code == 200
+    assert response.status_code == 201
     response_data = json.loads(response.data)
     pytest.global_alternative_identifier_id = response_data[0]["id"]
 
@@ -110,7 +110,13 @@ def test_delete_alternative_identifier(_logged_in_client):
         f"/study/{study_id}/dataset/{dataset_id}/metadata/alternative-identifier/{identifier_id}"
     )
 
-    assert response.status_code == 200
+    assert response.status_code == 204
+    response_get = _logged_in_client.get(
+        f"/study/{study_id}/dataset/{dataset_id}/metadata/alternative-identifier"
+    )
+    assert response_get.status_code == 200
+
+    assert len(json.loads(response_get.data)) == 0
 
 
 # ------------------- CONSENT METADATA ------------------- #
@@ -197,7 +203,7 @@ def test_post_dataset_contributor_metadata(_logged_in_client):
         ],
     )
 
-    assert response.status_code == 200
+    assert response.status_code == 201
     response_data = json.loads(response.data)
     pytest.global_dataset_contributor_id = response_data[0]["id"]
 
@@ -246,7 +252,14 @@ def test_delete_dataset_contributor_metadata(_logged_in_client):
         f"/study/{study_id}/dataset/{dataset_id}/metadata/contributor/{contributor_id}"
     )
 
-    assert response.status_code == 200
+    assert response.status_code == 204
+
+    response_get = _logged_in_client.get(
+        f"/study/{study_id}/dataset/{dataset_id}/metadata/contributor"
+    )
+    assert response_get.status_code == 200
+
+    assert len(json.loads(response_get.data)) == 0
 
 
 # ------------------- CREATOR METADATA ------------------- #
@@ -296,7 +309,7 @@ def test_post_dataset_creator_metadata(_logged_in_client):
         ],
     )
 
-    assert response.status_code == 200
+    assert response.status_code == 201
     response_data = json.loads(response.data)
     pytest.global_dataset_creator_id = response_data[0]["id"]
 
@@ -327,7 +340,14 @@ def test_delete_dataset_creator_metadata(_logged_in_client):
         f"/study/{study_id}/dataset/{dataset_id}/metadata/creator/{creator_id}"
     )
 
-    assert response.status_code == 200
+    assert response.status_code == 204
+
+    response_get = _logged_in_client.get(
+        f"/study/{study_id}/dataset/{dataset_id}/metadata/creator"
+    )
+    assert response_get.status_code == 200
+
+    assert len(json.loads(response_get.data)) == 0
 
 
 # ------------------- DATE METADATA ------------------- #
@@ -363,7 +383,7 @@ def test_post_dataset_date_metadata(_logged_in_client):
         json=[{"date": 20210101, "type": "Type", "information": "Info"}],
     )
 
-    assert response.status_code == 200
+    assert response.status_code == 201
     response_data = json.loads(response.data)
     pytest.global_dataset_date_id = response_data[0]["id"]
 
@@ -387,7 +407,14 @@ def test_delete_dataset_date_metadata(_logged_in_client):
         f"/study/{study_id}/dataset/{dataset_id}/metadata/date/{date_id}"
     )
 
-    assert response.status_code == 200
+    assert response.status_code == 204
+
+    response_get = _logged_in_client.get(
+        f"/study/{study_id}/dataset/{dataset_id}/metadata/date"
+    )
+    assert response_get.status_code == 200
+
+    assert len(json.loads(response_get.data)) == 0
 
 
 # ------------------- DE-IDENTIFICATION LEVEL METADATA ------------------- #
@@ -480,7 +507,7 @@ def test_post_dataset_description_metadata(_logged_in_client):
         json=[{"description": "Description", "type": "Methods"}],
     )
 
-    assert response.status_code == 200
+    assert response.status_code == 201
     response_data = json.loads(response.data)
     pytest.global_dataset_description_id = response_data[0]["id"]
 
@@ -504,7 +531,15 @@ def test_delete_dataset_description_metadata(_logged_in_client):
         f"/study/{study_id}/dataset/{dataset_id}/metadata/description/{description_id}"
     )
 
-    assert response.status_code == 200
+    assert response.status_code == 204
+
+    response_get = _logged_in_client.get(
+        f"/study/{study_id}/dataset/{dataset_id}/metadata/description"
+    )
+    assert (
+        len(json.loads(response_get.data)) == 1
+        and json.loads(response_get.data)[0]["type"] == "Abstract"
+    )
 
 
 # ------------------- FUNDER METADATA ------------------- #
@@ -552,7 +587,7 @@ def test_post_dataset_funder_metadata(_logged_in_client):
         ],
     )
 
-    assert response.status_code == 200
+    assert response.status_code == 201
     response_data = json.loads(response.data)
     pytest.global_dataset_funder_id = response_data[0]["id"]
 
@@ -581,7 +616,13 @@ def test_delete_dataset_funder_metadata(_logged_in_client):
         f"/study/{study_id}/dataset/{dataset_id}/metadata/funder/{funder_id}"
     )
 
-    assert response.status_code == 200
+    assert response.status_code == 204
+
+    response_get = _logged_in_client.get(
+        f"/study/{study_id}/dataset/{dataset_id}/metadata/funder"
+    )
+    assert response_get.status_code == 200
+    assert len(json.loads(response_get.data)) == 0
 
 
 # ------------------- OTHER METADATA ------------------- #
@@ -720,9 +761,8 @@ def test_put_dataset_record_keys_metadata(_logged_in_client):
         json={"type": "Record Type", "details": "Details for Record Keys"},
     )
 
-    assert response.status_code == 201
+    assert response.status_code == 200
     response_data = json.loads(response.data)
-
     assert response_data["type"] == "Record Type"
     assert response_data["details"] == "Details for Record Keys"
 
@@ -787,7 +827,10 @@ def test_post_dataset_related_item_metadata(_logged_in_client):
                 "publication_year": 2013,
                 "publisher": "Publisher",
                 "relation_type": "Relation Type",
-                "titles": [{"title": "Title", "type": "MainTitle"}],
+                "titles": [
+                    {"title": "Title", "type": "MainTitle"},
+                    {"title": "Title", "type": "Subtitle"},
+                ],
                 "type": "Type",
                 "volume": "Volume",
             }
@@ -829,6 +872,8 @@ def test_post_dataset_related_item_metadata(_logged_in_client):
     assert response_data[0]["relation_type"] == "Relation Type"
     assert response_data[0]["titles"][0]["title"] == "Title"
     assert response_data[0]["titles"][0]["type"] == "MainTitle"
+    assert response_data[0]["titles"][1]["title"] == "Title"
+    assert response_data[0]["titles"][1]["type"] == "Subtitle"
     assert response_data[0]["type"] == "Type"
     assert response_data[0]["volume"] == "Volume"
 
@@ -851,7 +896,13 @@ def test_delete_dataset_related_item_contributor_metadata(_logged_in_client):
         f"/study/{study_id}/dataset/{dataset_id}/metadata/related-item/{related_item_id}/contributor/{contributor_id}"
     )
 
-    assert response.status_code == 200
+    assert response.status_code == 204
+
+    response_get = _logged_in_client.get(
+        f"/study/{study_id}/dataset/{dataset_id}/metadata/related-item"
+    )
+    assert response_get.status_code == 200
+    assert len(json.loads(response_get.data)[0]["contributors"]) == 0
 
 
 def test_delete_dataset_related_item_creator_metadata(_logged_in_client):
@@ -872,7 +923,13 @@ def test_delete_dataset_related_item_creator_metadata(_logged_in_client):
         f"/study/{study_id}/dataset/{dataset_id}/metadata/related-item/{related_item_id}/creator/{creator_id}"
     )
 
-    assert response.status_code == 200
+    assert response.status_code == 204
+
+    response_get = _logged_in_client.get(
+        f"/study/{study_id}/dataset/{dataset_id}/metadata/related-item"
+    )
+
+    assert len(json.loads(response_get.data)[0]["creators"]) == 0
 
 
 def test_delete_dataset_related_item_identifier_metadata(_logged_in_client):
@@ -893,7 +950,14 @@ def test_delete_dataset_related_item_identifier_metadata(_logged_in_client):
         f"/study/{study_id}/dataset/{dataset_id}/metadata/related-item/{related_item_id}/identifier/{identifier_id}"
     )
 
-    assert response.status_code == 200
+    assert response.status_code == 204
+
+    response_get = _logged_in_client.get(
+        f"/study/{study_id}/dataset/{dataset_id}/metadata/related-item"
+    )
+
+    assert response_get.status_code == 200
+    assert len(json.loads(response_get.data)[0]["identifiers"]) == 0
 
 
 def test_delete_dataset_related_item_title_metadata(_logged_in_client):
@@ -907,14 +971,35 @@ def test_delete_dataset_related_item_title_metadata(_logged_in_client):
     study_id = pytest.global_study_id["id"]  # type: ignore
     dataset_id = pytest.global_dataset_id
     related_item_id = pytest.global_dataset_related_item_id
-    title_id = pytest.global_dataset_related_item_title_id
-
+    # title_id = pytest.global_dataset_related_item_title_id
     # pylint: disable=line-too-long
-    response = _logged_in_client.delete(
-        f"/study/{study_id}/dataset/{dataset_id}/metadata/related-item/{related_item_id}/title/{title_id}"
+    response_get = _logged_in_client.get(
+        f"/study/{study_id}/dataset/{dataset_id}/metadata/related-item"
     )
 
-    assert response.status_code == 200
+    # titles_to_delete = [
+    #     # f"/study/{study_id}/dataset/{dataset_id}/metadata/related-item/{related_item_id}/title/{title_id}"
+    #     for i in json.loads(response_get.data)[0]["titles"]
+    #     if i["type"] != "MainTitle"
+    # ]
+
+    for i in json.loads(response_get.data)[0]["titles"]:
+        if i["type"] != "MainTitle":
+            t_id = i["id"]
+            response = _logged_in_client.delete(
+                f"/study/{study_id}/dataset/{dataset_id}/metadata/related-item/{related_item_id}/title/{t_id}"
+            )
+            assert response.status_code == 204
+
+    response_get = _logged_in_client.get(
+        f"/study/{study_id}/dataset/{dataset_id}/metadata/related-item"
+    )
+
+    assert response_get.status_code == 200
+    assert (
+        len(json.loads(response_get.data)[0]["titles"]) == 1
+        and json.loads(response_get.data)[0]["titles"][0]["type"] == "MainTitle"
+    )
 
 
 def test_delete_dataset_related_item_metadata(_logged_in_client):
@@ -933,7 +1018,13 @@ def test_delete_dataset_related_item_metadata(_logged_in_client):
         f"/study/{study_id}/dataset/{dataset_id}/metadata/related-item/{related_item_id}"
     )
 
-    assert response.status_code == 200
+    assert response.status_code == 204
+    response_get = _logged_in_client.get(
+        f"/study/{study_id}/dataset/{dataset_id}/metadata/related-item"
+    )
+
+    assert response_get.status_code == 200
+    assert len(json.loads(response_get.data)) == 0
 
 
 # ------------------- RIGHTS METADATA ------------------- #
@@ -978,7 +1069,7 @@ def test_post_dataset_rights_metadata(_logged_in_client):
         ],
     )
 
-    assert response.status_code == 200
+    assert response.status_code == 201
     response_data = json.loads(response.data)
     pytest.global_dataset_rights_id = response_data[0]["id"]
 
@@ -1004,7 +1095,14 @@ def test_delete_dataset_rights_metadata(_logged_in_client):
         f"/study/{study_id}/dataset/{dataset_id}/metadata/rights/{rights_id}"
     )
 
-    assert response.status_code == 200
+    assert response.status_code == 204
+
+    response_get = _logged_in_client.get(
+        f"/study/{study_id}/dataset/{dataset_id}/metadata/rights"
+    )
+    assert response_get.status_code == 200
+
+    assert len(json.loads(response_get.data)) == 0
 
 
 # ------------------- SUBJECTS METADATA ------------------- #
@@ -1050,7 +1148,7 @@ def test_post_dataset_subjects_metadata(_logged_in_client):
         ],
     )
 
-    assert response.status_code == 200
+    assert response.status_code == 201
     response_data = json.loads(response.data)
     pytest.global_dataset_subject_id = response_data[0]["id"]
 
@@ -1076,7 +1174,14 @@ def test_delete_dataset_subject_metadata(_logged_in_client):
         f"/study/{study_id}/dataset/{dataset_id}/metadata/subject/{subject_id}"
     )
 
-    assert response.status_code == 200
+    assert response.status_code == 204
+
+    response_get = _logged_in_client.get(
+        f"/study/{study_id}/dataset/{dataset_id}/metadata/subject"
+    )
+    assert response_get.status_code == 200
+
+    assert len(json.loads(response_get.data)) == 0
 
 
 # ------------------- TITLE METADATA ------------------- #
@@ -1114,7 +1219,7 @@ def test_post_dataset_title_metadata(_logged_in_client):
         json=[{"title": "Title", "type": "Subtitle"}],
     )
 
-    assert response.status_code == 200
+    assert response.status_code == 201
     response_data = json.loads(response.data)
     pytest.global_dataset_title_id = response_data[0]["id"]
 
@@ -1138,4 +1243,13 @@ def test_delete_dataset_title_metadata(_logged_in_client):
         f"/study/{study_id}/dataset/{dataset_id}/metadata/title/{title_id}"
     )
 
-    assert response.status_code == 200
+    assert response.status_code == 204
+
+    response_get = _logged_in_client.get(
+        f"/study/{study_id}/dataset/{dataset_id}/metadata/title"
+    )
+    assert response_get.status_code == 200
+    assert (
+        len(json.loads(response_get.data)) == 1
+        and json.loads(response_get.data)[0]["type"] == "MainTitle"
+    )
